Adding helper methods and attributes to support smoke testing
Adding helper methods to collect the names of the images that should have been created by packer, and a statement in the _publish recipe to save those image names to the node object. This will be used by Delivery in the Smoke phase to see if the images were actually created.